DumbTracer: Accurately determine the crashing instruction.
The old implementation assumes that the overflowing payload is on the
stack starting at the stack pointer at the beginning of the crashing
basic block. Unfortunately, this does not always hold. This commit
forces DumbTracer to step until the crashing instruction and find
overflowing payload on the stack at that point.